+++ packages/common/src/types/openIdIdentity.ts
@@ -20,7 +20,8 @@ export type UpdateOpenIdentity = Pick<
     'subject' | 'issuer' | 'email' | 'issuerType' | 'refreshToken'
 >;
 
-export type OpenIdIdentity = CreateOpenIdIdentity & {
+export type OpenIdIdentity = Omit<CreateOpenIdIdentity, 'userId'> & {
+    userUuid: string;
     createdAt: Date;
 };