+++ packages/e2e/cypress/e2e/app/dates.cy.ts
@@ -1,6 +1,5 @@
 import { SEED_PROJECT } from '@lightdash/common';
 import dayjs = require('dayjs');
-import moment = require('moment');
 
 function leadingZero(value: string | number) {
     return `0${value}`.slice(-2);
@@ -12,6 +11,21 @@ function getLocalISOString(date: Date) {
     return `${date.getFullYear()}-${getFullMonth(date)}-${leadingZero(
         date.getDate(),
     )}`;
+}
+function getHourOffset() {
+    const now = new Date('1 January, 2000'); // set specific date to avoid daylight savings
+    const regex = /([+-]\d{4})/;
+    const match = now.toTimeString().match(regex);
+
+    // Check if we found a match
+    if (match) {
+        return match[1]; // eg: +0200
+    }
+    return null; // Return null if no match was found
+}
+
+function getLocalTime(datetime: string) {
+    return dayjs(`${datetime} UTC`).format('YYYY-MM-DD HH:mm:ss');
 }
 
 describe('Date tests', () => {
@@ -27,15 +41,34 @@ describe('Date tests', () => {
         switch (timezone) {
             case 'UTC':
                 expect(offset, errorMessage).to.be.equal(0);
+                expect(getHourOffset()).to.be.equal('+0000');
+                expect(getLocalTime('2020-08-11 22:58:00')).to.be.equal(
+                    '2020-08-11 22:58:00',
+                );
+
                 break;
             case 'Europe/Madrid':
                 expect(offset, errorMessage).to.be.equal(-60);
+                expect(getHourOffset()).to.be.equal('+0100');
+                expect(getLocalTime('2020-08-11 22:58:00')).to.be.equal(
+                    '2020-08-12 00:58:00',
+                );
                 break;
             case 'America/New_York':
                 expect(offset, errorMessage).to.be.equal(300);
+                expect(getHourOffset()).to.be.equal('-0500');
+                expect(getLocalTime('2020-08-11 22:58:00')).to.be.equal(
+                    '2020-08-11 18:58:00',
+                );
+
                 break;
             case 'Asia/Tokyo':
                 expect(offset, errorMessage).to.be.equal(-540);
+                expect(getHourOffset()).to.be.equal('+0900');
+                expect(getLocalTime('2020-08-11 22:58:00')).to.be.equal(
+                    '2020-08-12 07:58:00',
+                );
+
                 break;
             default:
                 throw new Error(
@@ -126,144 +159,169 @@ describe('Date tests', () => {
         });
     });
 
-    it.skip('Should use UTC dates', () => {
-        const exploreStateUrlParams = `?create_saved_chart_version=%7B%22tableName%22%3A%22orders%22%2C%22metricQuery%22%3A%7B%22dimensions%22%3A%5B%22orders_status%22%2C%22customers_created_raw%22%5D%2C%22metrics%22%3A%5B%22orders_average_order_size%22%5D%2C%22filters%22%3A%7B%22dimensions%22%3A%7B%22id%22%3A%22927e8fc4-4a41-4972-8d15-57cb2060a1d2%22%2C%22and%22%3A%5B%7B%22id%22%3A%228cf33dc8-d62a-41fa-85c8-4078e028bd60%22%2C%22target%22%3A%7B%22fieldId%22%3A%22customers_created_raw%22%7D%2C%22operator%22%3A%22lessThan%22%2C%22values%22%3A%5B%222022-07-11T14%3A23%3A11.302Z%22%5D%7D%5D%7D%7D%2C%22sorts%22%3A%5B%7B%22fieldId%22%3A%22customers_created_raw%22%2C%22descending%22%3Atrue%7D%5D%2C%22limit%22%3A1%2C%22tableCalculations%22%3A%5B%5D%2C%22additionalMetrics%22%3A%5B%5D%7D%2C%22pivotConfig%22%3A%7B%22columns%22%3A%5B%22customers_created_raw%22%5D%7D%2C%22tableConfig%22%3A%7B%22columnOrder%22%3A%5B%22orders_status%22%2C%22customers_created_raw%22%2C%22orders_average_order_size%22%5D%7D%2C%22chartConfig%22%3A%7B%22type%22%3A%22cartesian%22%2C%22config%22%3A%7B%22layout%22%3A%7B%22xField%22%3A%22orders_status%22%2C%22yField%22%3A%5B%22orders_average_order_size%22%5D%7D%2C%22eChartsConfig%22%3A%7B%22series%22%3A%5B%7B%22type%22%3A%22bar%22%2C%22encode%22%3A%7B%22xRef%22%3A%7B%22field%22%3A%22orders_status%22%7D%2C%22yRef%22%3A%7B%22field%22%3A%22orders_average_order_size%22%2C%22pivotValues%22%3A%5B%7B%22field%22%3A%22customers_created_raw%22%2C%22value%22%3A%222017-02-11T03%3A00%3A00.000Z%22%7D%5D%7D%7D%7D%5D%2C%22legend%22%3A%7B%22show%22%3Atrue%2C%22type%22%3A%22plain%22%2C%22orient%22%3A%22horizontal%22%7D%7D%7D%7D%7D`;
+    it('Should use UTC dates', () => {
+        const exploreStateUrlParams = `?create_saved_chart_version={"tableName"%3A"events"%2C"metricQuery"%3A{"exploreName"%3A""%2C"dimensions"%3A["events_timestamp_tz_raw"]%2C"metrics"%3A["events_count"]%2C"filters"%3A{"dimensions"%3A{"id"%3A"3b565490-87c5-4996-a42b-ff0640bb18cd"%2C"and"%3A[{"id"%3A"be863f3c-5807-48c5-9b6f-2e8445610280"%2C"target"%3A{"fieldId"%3A"events_timestamp_tz_raw"}%2C"operator"%3A"equals"%2C"values"%3A["2020-08-12T00%3A58%3A00%2B02%3A00"]}]}}%2C"sorts"%3A[{"fieldId"%3A"events_timestamp_tz_raw"%2C"descending"%3Atrue}]%2C"limit"%3A500%2C"tableCalculations"%3A[]%2C"additionalMetrics"%3A[]}%2C"tableConfig"%3A{"columnOrder"%3A["events_timestamp_tz_raw"%2C"events_count"]}%2C"chartConfig"%3A{"type"%3A"cartesian"%2C"config"%3A{"layout"%3A{"xField"%3A"events_timestamp_tz_raw"%2C"yField"%3A["events_count"]}%2C"eChartsConfig"%3A{"series"%3A[{"type"%3A"bar"%2C"yAxisIndex"%3A0%2C"encode"%3A{"xRef"%3A{"field"%3A"events_timestamp_tz_raw"}%2C"yRef"%3A{"field"%3A"events_count"}}}]}}}}`;
         cy.visit(
-            `/projects/${SEED_PROJECT.project_uuid}/tables/orders${exploreStateUrlParams}`,
+            `/projects/${SEED_PROJECT.project_uuid}/tables/events${exploreStateUrlParams}`,
         );
-
-        cy.findByTestId('Filters-card-expand').click();
+        cy.contains('SQL');
         cy.findAllByText('Loading chart').should('have.length', 0);
-        cy.findByDisplayValue('2022-07-11, 14:23:11:000');
-        cy.get('svg g text').contains('2017-02-11, 03:00:00:000 (+00:00)');
-        cy.get('tbody td').contains('2017-02-11, 03:00:00:000 (+00:00)');
+
+        cy.contains('1 active filter');
+        cy.findByTestId('Filters-card-expand').click();
+        cy.contains('11 Aug 2020 22:58:00'); // Filter in UTC
+
+        cy.contains(`2020-08-11, 22:58:00:000 (+00:00)`); // Data in results, this comes from the server, so depends on the server timezone
+
+        // Time sensitive fields in localtime
+        const timezone = Cypress.env('TZ');
+        switch (timezone) {
+            case 'UTC':
+                cy.contains('2020-08-11 22:58:00'); // Filter in localtime //Timezone sensitive
+                break;
+            case 'Europe/Madrid':
+                cy.contains('2020-08-12 00:58:00'); // Filter in localtime //Timezone sensitive
+                break;
+            case 'America/New_York':
+                cy.contains('2020-08-11 18:58:00'); // Filter in localtime //Timezone sensitive
+                break;
+            case 'Asia/Tokyo':
+                cy.contains('2020-08-12 07:58:00'); // Filter in localtime //Timezone sensitive
+                break;
+            default:
+                throw new Error(
+                    `Invalid timezone (${timezone}) to test, please add CYPRESS_TZ to your env with a valid timezone`,
+                );
+        }
+
         cy.findByTestId('SQL-card-expand').click();
-        cy.get('code')
-            .invoke('text')
-            .should(
-                'include',
-                '("customers".created) < (\'2022-07-11 14:23:11\')',
-            );
+        cy.contains(`("events".timestamp_tz) = ('2020-08-11 22:58:00')`); // SQL
+
+        // TODO tooltip in charts
+        // cy.get('svg g').trigger('mouseover')
+        // cy.contains('2020-08-11 22:58:00');
     });
 
-    it.skip('Should filter by date on results table', () => {
+    it('Should filter by date on results table', () => {
+        // This test should not be timezone sensitive
         const exploreStateUrlParams = `?create_saved_chart_version=%7B%22tableName%22%3A%22orders%22%2C%22metricQuery%22%3A%7B%22dimensions%22%3A%5B%22orders_order_date_day%22%2C%22orders_order_date_week%22%2C%22orders_order_date_month%22%2C%22orders_order_date_year%22%5D%2C%22metrics%22%3A%5B%5D%2C%22filters%22%3A%7B%7D%2C%22sorts%22%3A%5B%7B%22fieldId%22%3A%22orders_order_date_day%22%2C%22descending%22%3Atrue%7D%5D%2C%22limit%22%3A1%2C%22tableCalculations%22%3A%5B%5D%2C%22additionalMetrics%22%3A%5B%5D%7D%2C%22tableConfig%22%3A%7B%22columnOrder%22%3A%5B%22orders_order_date_day%22%2C%22orders_order_date_week%22%2C%22orders_order_date_month%22%2C%22orders_order_date_year%22%5D%7D%2C%22chartConfig%22%3A%7B%22type%22%3A%22cartesian%22%2C%22config%22%3A%7B%22layout%22%3A%7B%22xField%22%3A%22orders_order_date_day%22%2C%22yField%22%3A%5B%22orders_order_date_week%22%5D%7D%2C%22eChartsConfig%22%3A%7B%22series%22%3A%5B%7B%22encode%22%3A%7B%22xRef%22%3A%7B%22field%22%3A%22orders_order_date_day%22%7D%2C%22yRef%22%3A%7B%22field%22%3A%22orders_order_date_week%22%7D%7D%2C%22type%22%3A%22bar%22%7D%5D%7D%7D%7D%7D`;
         cy.visit(
             `/projects/${SEED_PROJECT.project_uuid}/tables/orders${exploreStateUrlParams}`,
         );
+        cy.contains('SQL');
+        cy.findAllByText('Loading chart').should('have.length', 0);
 
         cy.findByTestId('Filters-card-expand').click();
         cy.findByTestId('SQL-card-expand').click();
+        cy.findByTestId('Chart-card-expand').click(); // Close chart
 
-        cy.findAllByText('Loading chart').should('have.length', 0);
         // Filter by year
         cy.get('tbody > :nth-child(1) > :nth-child(5)').click();
-        cy.contains('Filter by "2018"').click();
-        cy.get('.bp4-numeric-input input').should('have.value', '2018');
-        cy.get('.bp4-code').contains(
+        cy.contains('Filter by 2018').click();
+        cy.get('.mantine-YearPickerInput-input').contains('2018');
+        cy.get('.mantine-Prism-code').contains(
             `(DATE_TRUNC('YEAR', "orders".order_date)) = ('2018-01-01')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
 
         // Filter by month
         cy.get('tbody > :nth-child(1) > :nth-child(4)').click();
-        cy.contains('Filter by "2018-04"').click();
-        cy.get('select option[label="January"]')
-            .parent('select')
-            .should('have.value', '3');
-        cy.get('.bp4-numeric-input input').should('have.value', '2018');
-        cy.get('.bp4-code').contains(
+        cy.contains('Filter by 2018-04').click();
+
+        cy.get('.mantine-MonthPickerInput-input').contains('April 2018');
+        cy.get('.mantine-Prism-code').contains(
             `(DATE_TRUNC('MONTH', "orders".order_date)) = ('2018-04-01')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
 
         // Filter by week
         cy.get('tbody > :nth-child(1) > :nth-child(3)').click();
-        cy.contains('Filter by "2018-04-09"').click();
-        cy.get('.bp4-date-input input').should('have.value', '2018-04-09');
-        cy.get('.bp4-code').contains(
+        cy.contains('Filter by 2018-04-09').click();
+        cy.get('.mantine-DateInput-input').should(
+            'have.value',
+            'April 9, 2018',
+        );
+        cy.get('.mantine-Prism-code').contains(
             `(DATE_TRUNC('WEEK', "orders".order_date)) = ('2018-04-09')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
 
         // Filter by day
         cy.get('tbody > :nth-child(1) > :nth-child(2)').click();
-        cy.contains('Filter by "2018-04-09"').click();
-        cy.get('.bp4-date-input input').should('have.value', '2018-04-09');
-        cy.get('.bp4-code').contains(
+        cy.contains('Filter by 2018-04-09').click();
+        cy.get('.mantine-DateInput-input').should(
+            'have.value',
+            'April 9, 2018',
+        );
+        cy.get('.mantine-Prism-code').contains(
             `(DATE_TRUNC('DAY', "orders".order_date)) = ('2018-04-09')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
     });
 
-    it.skip('Should filter by datetimes on results table', () => {
+    it('Should filter by datetimes on results table', () => {
         const exploreStateUrlParams = `?create_saved_chart_version={"tableName"%3A"events"%2C"metricQuery"%3A{"dimensions"%3A["events_timestamp_tz_raw"%2C"events_timestamp_tz_millisecond"%2C"events_timestamp_tz_second"%2C"events_timestamp_tz_minute"%2C"events_timestamp_tz_hour"]%2C"metrics"%3A[]%2C"filters"%3A{}%2C"sorts"%3A[{"fieldId"%3A"events_timestamp_tz_raw"%2C"descending"%3Atrue}]%2C"limit"%3A1%2C"tableCalculations"%3A[]%2C"additionalMetrics"%3A[]}%2C"tableConfig"%3A{"columnOrder"%3A["events_timestamp_tz_raw"%2C"events_timestamp_tz_millisecond"%2C"events_timestamp_tz_second"%2C"events_timestamp_tz_minute"%2C"events_timestamp_tz_hour"]}%2C"chartConfig"%3A{"type"%3A"cartesian"%2C"config"%3A{"layout"%3A{"xField"%3A"events_timestamp_tz_raw"%2C"yField"%3A["events_timestamp_tz_millisecond"]}%2C"eChartsConfig"%3A{"series"%3A[{"encode"%3A{"xRef"%3A{"field"%3A"events_timestamp_tz_raw"}%2C"yRef"%3A{"field"%3A"events_timestamp_tz_millisecond"}}%2C"type"%3A"bar"}]}}}}`;
         cy.visit(
             `/projects/${SEED_PROJECT.project_uuid}/tables/events${exploreStateUrlParams}`,
         );
+        cy.contains('SQL');
+        cy.findAllByText('Loading chart').should('have.length', 0);
 
         cy.findByTestId('Filters-card-expand').click();
         cy.findByTestId('SQL-card-expand').click();
         cy.findByTestId('Chart-card-expand').click(); // Close chart
 
-        cy.findAllByText('Loading chart').should('have.length', 0);
         // Filter by raw
         cy.get('tbody > :nth-child(1) > :nth-child(2)').click();
-        cy.contains('Filter by "2019-12-02, 00:00:00:000 (+00:00)').click();
-        cy.get('.bp4-date-input input').should(
-            'have.value',
-            '2019-12-02, 00:00:00:000',
+        cy.contains('Filter by 2020-08-11, 23:44:00:000 (+00:00)').click(); // Server Timezone sensitive
+        cy.get('.mantine-DateTimePicker-input').contains(
+            getLocalTime('2020-08-11 23:44:00'), // Timezone sensitive
         );
-        cy.get('.bp4-code').contains(
-            `("events".timestamp_tz) = ('2019-12-02 00:00:00')`,
+        cy.get('.mantine-Prism-code').contains(
+            `("events".timestamp_tz) = ('2020-08-11 23:44:00')`,
         );
 
-        cy.scrollTo('top');
         cy.get('.tabler-icon-x').click({ multiple: true, force: true });
         // Filter by Milisecond
         // FIXME: selecting a different cell is not working
         cy.get('tbody > :nth-child(1) > :nth-child(3)').click();
-        cy.contains('Filter by "2019-12-02, 00:00:00:000 (+00:00)').click();
-        cy.get('.bp4-date-input input').should(
-            'have.value',
-            '2019-12-02, 00:00:00:000',
+        cy.contains('Filter by 2020-08-11, 23:44:00:000 (+00:00)').click(); // Server Timezone sensitive
+        cy.get('.mantine-DateTimePicker-input').contains(
+            getLocalTime('2020-08-11 23:44:00'), // Timezone sensitive
         );
-        cy.get('.bp4-code').contains(
-            `(DATE_TRUNC('MILLISECOND', "events".timestamp_tz)) = ('2019-12-02 00:00:00')`, // Known Milisecond limitation
+        cy.get('.mantine-Prism-code').contains(
+            `(DATE_TRUNC('MILLISECOND', "events".timestamp_tz)) = ('2020-08-11 23:44:00')`, // Known Milisecond limitation
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
 
         // Filter by Second
         cy.get('tbody > :nth-child(1) > :nth-child(4)').click();
-        cy.contains('Filter by "2019-12-02, 00:00:00 (+00:00)').click();
-        cy.get('.bp4-date-input input').should(
-            'have.value',
-            '2019-12-02, 00:00:00:000',
+        cy.contains('Filter by 2020-08-11, 23:44:00 (+00:00)').click(); // Server Timezone sensitive
+        cy.get('.mantine-DateTimePicker-input').contains(
+            getLocalTime('2020-08-11 23:44:00'), // Timezone sensitive
         );
-        cy.get('.bp4-code').contains(
-            `(DATE_TRUNC('SECOND', "events".timestamp_tz)) = ('2019-12-02 00:00:00')`,
+        cy.get('.mantine-Prism-code').contains(
+            `(DATE_TRUNC('SECOND', "events".timestamp_tz)) = ('2020-08-11 23:44:00')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
         // Filter by Minute
         cy.get('tbody > :nth-child(1) > :nth-child(5)').click();
-        cy.contains('Filter by "2019-12-02, 00:00 (+00:00)').click();
-        cy.get('.bp4-date-input input').should(
-            'have.value',
-            '2019-12-02, 00:00:00:000',
+        cy.contains('Filter by 2020-08-11, 23:44 (+00:00)').click(); // Server Timezone sensitive
+        cy.get('.mantine-DateTimePicker-input').contains(
+            getLocalTime('2020-08-11 23:44:00'), // Timezone sensitive
         );
-        cy.get('.bp4-code').contains(
-            `(DATE_TRUNC('MINUTE', "events".timestamp_tz)) = ('2019-12-02 00:00:00')`,
+        cy.get('.mantine-Prism-code').contains(
+            `(DATE_TRUNC('MINUTE', "events".timestamp_tz)) = ('2020-08-11 23:44:00')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
 
         // Filter by Hour
         cy.get('tbody > :nth-child(1) > :nth-child(6)').click();
-        cy.contains('Filter by "2020-08-11, 23 (+00:00)').click();
-        cy.get('.bp4-date-input input').should(
-            'have.value',
-            '2020-08-11, 23:00:00:000',
+        cy.contains('Filter by 2020-08-11, 23 (+00:00)').click(); // Server Timezone sensitive
+        cy.get('.mantine-DateTimePicker-input').contains(
+            getLocalTime('2020-08-11 23:00:00'), // Timezone sensitive
         );
-        cy.get('.bp4-code').contains(
+        cy.get('.mantine-Prism-code').contains(
             `(DATE_TRUNC('HOUR', "events".timestamp_tz)) = ('2020-08-11 23:00:00')`,
         );
         cy.get('.tabler-icon-x').click({ multiple: true });
@@ -308,9 +366,9 @@ describe('Date tests', () => {
         cy.contains('Add filter').click();
         cy.contains('Created month').click();
 
-        cy.contains('button', moment().format('MMMM YYYY')).click();
+        cy.contains('button', dayjs().format('MMMM YYYY')).click();
         cy.findByRole('dialog').within(() => {
-            cy.contains('button', moment().format('YYYY')).click();
+            cy.contains('button', dayjs().format('YYYY')).click();
             cy.get('button').find('[data-previous="true"]').click();
             cy.contains('button', 2017).click();
             cy.contains('button', 'Aug').click();
@@ -390,7 +448,7 @@ describe('Date tests', () => {
     });
 
     it('Should filter by date on dimension', () => {
-        const now = moment();
+        const now = dayjs();
         const exploreStateUrlParams = `?create_saved_chart_version=%7B%22tableName%22%3A%22orders%22%2C%22metricQuery%22%3A%7B%22dimensions%22%3A%5B%22orders_order_date_day%22%2C%22orders_order_date_week%22%2C%22orders_order_date_month%22%2C%22orders_order_date_year%22%5D%2C%22metrics%22%3A%5B%5D%2C%22filters%22%3A%7B%7D%2C%22sorts%22%3A%5B%7B%22fieldId%22%3A%22orders_order_date_day%22%2C%22descending%22%3Atrue%7D%5D%2C%22limit%22%3A1%2C%22tableCalculations%22%3A%5B%5D%2C%22additionalMetrics%22%3A%5B%5D%7D%2C%22tableConfig%22%3A%7B%22columnOrder%22%3A%5B%22orders_order_date_day%22%2C%22orders_order_date_week%22%2C%22orders_order_date_month%22%2C%22orders_order_date_year%22%5D%7D%2C%22chartConfig%22%3A%7B%22type%22%3A%22cartesian%22%2C%22config%22%3A%7B%22layout%22%3A%7B%22xField%22%3A%22orders_order_date_day%22%2C%22yField%22%3A%5B%22orders_order_date_week%22%5D%7D%2C%22eChartsConfig%22%3A%7B%22series%22%3A%5B%7B%22encode%22%3A%7B%22xRef%22%3A%7B%22field%22%3A%22orders_order_date_day%22%7D%2C%22yRef%22%3A%7B%22field%22%3A%22orders_order_date_week%22%7D%7D%2C%22type%22%3A%22bar%22%7D%5D%7D%7D%7D%7D`;
         cy.visit(
             `/projects/${SEED_PROJECT.project_uuid}/tables/orders${exploreStateUrlParams}`,
@@ -489,8 +547,8 @@ describe('Date tests', () => {
         cy.findAllByText('Loading chart').should('have.length', 0);
 
         const checkDatetime = ($value, sqlFilter) => {
-            const now = moment();
-            const aSecondBefore = moment().subtract(1, 'seconds'); // Fix millisecond race condition
+            const now = dayjs();
+            const aSecondBefore = dayjs().subtract(1, 'seconds'); // Fix millisecond race condition
             const dateString = $value?.val();
             const inputDatetimeFormat = 'YYYY-MM-DD, HH:mm:ss:000';
             const expectedDatetimes = [
@@ -498,8 +556,8 @@ describe('Date tests', () => {
                 aSecondBefore.format(inputDatetimeFormat),
             ];
             expect(dateString).to.be.oneOf(expectedDatetimes);
-            cy.get('.bp4-code').contains(
-                `(${sqlFilter}) = ('${moment(dateString).format(
+            cy.get('.mantine-Prism-code').contains(
+                `(${sqlFilter}) = ('${dayjs(dateString).format(
                     'YYYY-MM-DD HH:mm:ss',
                 )}')`,
             );